+++ packages/vscode-graphql-syntax/tests/graphql-grammar.spec.ts
@@ -1,4 +1,3 @@
-import { describe, it, expect } from 'vitest';
 import { tokenizeFile } from './utilities';
 
 describe('source.graphql grammar', () => {
