+++ packages/vscode-graphql-syntax/vite.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    globals: true,
+    setupFiles: ['./serializer.ts'],
+  },
+});
